+++ apps/condo/domains/common/components/EmptyIcon.tsx
