+++ apps/condo/domains/common/components/GraphQlSearchInput.tsx
