+++ apps/condo/domains/common/components/Logo.tsx
