+++ apps/condo/domains/common/components/Tooltip.tsx
@@ -1,6 +1,6 @@
 import { Tooltip as DefaultTooltip, TooltipProps, Typography } from 'antd'
 import React from 'react'
-import { colors } from '../constants/style'
+import { colors } from '../../../constants/style'
 
 export const Tooltip: React.FC<TooltipProps> = (props) => (
     <DefaultTooltip
