+++ apps/condo/domains/common/utils/codegeneration/generate.hooks.ts
@@ -11,6 +11,10 @@ interface Refetch<Q> {
     (variables?: Q): Promise<any>
 }
 
+interface FetchMore<Q> {
+    (variables?: Q): Promise<any>
+}
+
 interface IHookConverters<GQL, GQLInput, UI, UIForm> {
     convertToGQLInput: (state: UIForm, item?: UI) => GQLInput
     convertToUIState: (item: GQL) => UI
@@ -26,7 +30,7 @@ interface IHookResult<UI, UIForm, Q, TData = any, TVariables = OperationVariable
 }
 
 export function generateReactHooks<GQL, GQLInput, UIForm, UI, Q> (gql, { convertToGQLInput, convertToUIState }: IHookConverters<GQL, GQLInput, UI, UIForm>): IHookResult<UI, UIForm, Q> {
-    
+
     function useObject (variables: Q, options?: QueryHookOptions<{ objs?: GQL[], meta?: { count?: number } }, Q>) {
         const { loading, refetch, objs, count, error } = useObjects(variables, options)
         if (count && count > 1) throw new Error('Wrong query condition! return more then one result')
@@ -37,7 +41,7 @@ export function generateReactHooks<GQL, GQLInput, UIForm, UI, Q> (gql, { convert
         const intl = useIntl()
         const ServerErrorPleaseTryAgainLaterMsg = intl.formatMessage({ id: 'ServerErrorPleaseTryAgainLater' })
         const AccessErrorMsg = intl.formatMessage({ id: 'AccessError' })
-        
+
         const result = useQuery<{ objs?: GQL[], meta?: { count?: number } }, Q>(gql.GET_ALL_OBJS_WITH_COUNT_QUERY, {
             variables,
             notifyOnNetworkStatusChange: true,
