+++ apps/condo/domains/ticket/components/BaseTicketForm/ErrorsContainer.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import { useIntl } from '@core/next/intl'
 import { Col } from 'antd'
 import React from 'react'
-import { colors } from '../../constants/style'
+import { colors } from '../../../../constants/style'
 
 export const ErrorsWrapper = styled.div`
   display: inline-block;
