+++ apps/condo/domains/ticket/components/BaseTicketForm/index.tsx
@@ -5,15 +5,15 @@ import styled from '@emotion/styled'
 import { Checkbox, Col, Form, Input, Row, Typography } from 'antd'
 import React from 'react'
 import { ITicketFormState } from '@condo/domains/ticket/utils/clientSchema/Ticket'
-import { colors } from '../../constants/style'
-import { FormWithAction } from '../../containers/FormList'
-import { searchEmployee, searchProperty, searchTicketClassifier } from '../../utils/clientSchema/Ticket/search'
-import { FocusContainer } from '../FocusContainer'
-import { GraphQlSearchInput } from '../GraphQlSearchInput'
-import { LabelWithInfo } from '../LabelWithInfo'
-import { PhoneInput } from '../PhoneInput'
-import { UnitNameInput } from '../UnitNameInput'
-import { UserNameField } from '../UserNameField'
+import { colors } from '../../../../constants/style'
+import { FormWithAction } from '../../../../containers/FormList'
+import { searchEmployee, searchProperty, searchTicketClassifier } from '../../../../utils/clientSchema/Ticket/search'
+import { FocusContainer } from '@condo/domains/common/components/FocusContainer'
+import { GraphQlSearchInput } from '@condo/domains/common/components/GraphQlSearchInput'
+import { LabelWithInfo } from '@condo/domains/common/components/LabelWithInfo'
+import { PhoneInput } from '@condo/domains/common/components/PhoneInput'
+import { UnitNameInput } from '@condo/domains/common/components/UnitNameInput'
+import { UserNameField } from '@condo/domains/common/components/UserNameField'
 import { useTicketValidations } from './useTicketValidations'
 
 const LAYOUT = {
