+++ apps/condo/domains/ticket/components/TicketForm/UpdateTicketForm.tsx
@@ -3,10 +3,10 @@ import React, { useEffect } from 'react'
 import { useOrganization } from '@core/next/organization'
 import { Col, Form, Row, Typography, Space } from 'antd'
 import { useIntl } from '@core/next/intl'
-import { BaseTicketForm } from '../../components/BaseTicketForm'
-import { Button } from '../../components/Button'
-import { ErrorsContainer } from '../../components/BaseTicketForm/ErrorsContainer'
-import { FormResetButton } from '../../components/FormResetButton'
+import { BaseTicketForm } from '../BaseTicketForm'
+import { Button } from '@condo/domains/common/components/Button'
+import { ErrorsContainer } from '../BaseTicketForm/ErrorsContainer'
+import { FormResetButton } from '@condo/domains/common/components/FormResetButton'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import { Ticket } from '@condo/domains/ticket/utils/clientSchema'
